pyproject.toml
[project]
name = "qgis-osm-conflator"
version = "0.1.0"
description = "A tool with similar functionalities to JOSM for OpenStreetMap data conflation (merging)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "osm-login-python>=2.0.0",
    "requests>=2.32.5",
]

[tool.qgis-plugin-ci]
plugin_path = "osm_conflator"
github_organization_slug = "hotosm"
project_slug = "qgis-osm-conflator"

[dependency-groups]
dev = [
    "ruff>=0.14.1",
    "qgis-plugin-ci>=2.8.10",
]

[tool.ruff]
fix = true
line-length = 88
target-version = "py311"
exclude = [
    ".ruff_cache",
    "osm_conflator/__version__.py",
]
[tool.ruff.lint]
select = ["I", "E", "W", "D", "B", "F", "N", "Q"]
ignore = ["N999", "N802"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_provider = "uv"
version_files = [
    "pyproject.toml:version",
]
changelog_file = "CHANGELOG.md"
update_changelog_on_bump = true

[tool.codespell]
skip = "*CHANGELOG.md"
write-changes = true
